% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_path.R
\name{convert_path}
\alias{convert_path}
\title{Convert Path and Create Missing Directories}
\usage{
convert_path(path)
}
\arguments{
\item{path}{Character string. The raw path (use raw strings or escaped backslashes).}
}
\value{
The converted path with forward slashes.
}
\description{
Normalizes slashes and creates the folder structure if missing.
}
