% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_labels.R
\name{pull_labels}
\alias{pull_labels}
\title{Save Variable Names and Labels in a List to Restore Them Later}
\usage{
pull_labels(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A named list where names are variable names and values are the labels.
}
\description{
Extracts variable names and variable labels from a data frame and stores them in a named list, so that labels can be re-applied to the dataframe after wrangling with \code{pull_labels} without losing them (e.g., with dplyr-based operations).
}
