% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_labels.R
\name{restore_labels}
\alias{restore_labels}
\title{Restore Variable Labels}
\usage{
restore_labels(data, label_list, renamed = NULL)
}
\arguments{
\item{data}{A data frame to label.}

\item{label_list}{A named list of labels (generated by \code{pull_labels}).}

\item{renamed}{A named character vector mapping new names to old names (e.g. c("new_name" = "old_name")).}
}
\value{
The data frame with restored labels.
}
\description{
Re-applies variable labels that have been extracted from the data frame earlier. Supports mapping for renamed variables.
}
