% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_overview.R
\name{describe_vars}
\alias{describe_vars}
\title{Overview of Variable Names, Labels and Classes in a Dataset}
\usage{
describe_vars(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A tibble with columns: variable, class, label.
}
\description{
Returns a clean table of variable names, classes, and labels.
Replaces NULL labels with NA.
}
